requirements.txt
requests>=2.28.1,<3.0.0
pynacl>=1.5.0,<1.6.0
cryptography>=38.0.1,<39.0.0
pyjwt>=2.5.0,<2.6.0
urllib3>=1.26.12,<1.27.0
pydantic
toml